refactor(website): route structure
