+++ packages/website/src/App.tsx
@@ -1,20 +1,13 @@
-import React, { FC, Suspense } from 'react'
-import { useRoutes, useLocation } from 'react-router-dom'
+import React, { Suspense } from 'react'
+import { useRoutes } from 'react-router-dom'
 import pageRoutes from '~react-pages'
-import { isNoLayoutRoute } from './utils/route'
 import { UserProvider } from './hooks/use-user'
-import GlobalLayout from './components/GlobalLayout'
 
-const App: FC = () => {
-  const location = useLocation()
+const App = () => {
   return (
     <UserProvider>
       <Suspense fallback={null}>
-        {
-          isNoLayoutRoute(location.pathname)
-            ? useRoutes(pageRoutes)
-            : <GlobalLayout>{useRoutes(pageRoutes)}</GlobalLayout>
-        }
+        {useRoutes(pageRoutes)}
       </Suspense>
     </UserProvider>
   )
