+++ packages/website/src/components/GlobalLayout/index.tsx
@@ -1,9 +1,9 @@
-import React from 'react'
+import React, { PropsWithChildren } from 'react'
 import Footer from '../Footer'
 import Header from '../Header'
 import styles from './style.module.less'
 
-const GlobalLayout: React.FC = (props) => {
+const GlobalLayout: React.FC<PropsWithChildren> = (props) => {
   return (
     <div className={styles.container}>
       <Header />
