+++ packages/website/src/pages/index/group/[name]/__tests__/fixtures/boring.json
