+++ packages/website/src/pages/index/group/[name]/__tests__/index.spec.tsx
@@ -7,7 +7,7 @@ import Boring from './fixtures/boring.json'
 import RecentTopics from './fixtures/recent-topics.json'
 import { MemoryRouter, Route, Routes, useParams } from 'react-router-dom'
 import { GroupProfile, ResponseWithPagination, Topic } from '@bangumi/types/group'
-import GroupPage from '@bangumi/website/pages/group/[name]'
+import GroupPage from '@bangumi/website/pages/index/group/[name]'
 
 jest.mock('react-router-dom', () => {
   return {
