+++ packages/website/src/pages/index/group/[name]/__tests__/members.spec.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { RenderResult, waitFor } from '@testing-library/react'
-import { renderPage } from '../../../../utils/test-utils'
-import { server as mockServer } from '../../../../mocks/server'
+import { renderPage } from '../../../../../utils/test-utils'
+import { server as mockServer } from '../../../../../mocks/server'
 import { rest } from 'msw'
 import GroupMembers from '../members'
 import Boring from './fixtures/boring.json'
