+++ packages/website/src/pages/index/group/[name]/components/ClampableContent.tsx
