+++ packages/website/src/pages/index/group/components/GroupHeader.tsx
