+++ packages/website/src/pages/index/index.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+const HomePage = () => {
+  return <div>Hi! This is Home page</div>
+}
+
+export default HomePage
