+++ packages/website/src/pages/index/subject/[id]/components/WikiEditor/DetailEditor.tsx
