+++ packages/website/src/pages/login/index.module.less
@@ -10,17 +10,22 @@
   top: 296px;
   transform: translateX(-50%);
   display: flex;
+  gap: 20px;
   flex-direction: column;
   align-items: center;
 
-  > * {
+  >* {
     width: 100%;
-    margin-bottom: 20px;
   }
 
   .hcaptcha div {
+    height: 78px;
     text-align: center;
   }
+}
+
+.logo {
+  margin-bottom: 25px;
 }
 
 .buttonGroup {
@@ -31,6 +36,8 @@
 
 .button {
   width: 150px;
+  // height: 38px;
+  font-weight: 600;
 }
 
 .icon {
